requirements.txt
amqp==5.2.0
async-timeout==4.0.3
billiard==3.6.4.0
celery==5.2.7
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
kombu==5.4.2
prompt_toolkit==3.0.47
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.8
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
